requirements.txt
textual
textual-dev
flask
pytest
pytest-flask
pytest-raises
pytest-asyncio
pytest-cov
coverage
mysql-connector-python